Factor out BaseOcrModel, add docling-parse backend tests, fixes
